media: improve webcam failure handling
If initialization of the video source leads to an error (e.g. because
the webcam is being used by another process and can't be opened), don't
end the call but try to replace the failed source with videotestsrc and
continue.

For the user it is a better outcome to have audio and possibly one-way
video with the other person than not to be able to make a call at all.