proxy: don't use default proxy resolver when none should be used
It isn't guaranteed g_proxy_resolver_get_default() always returns
a resolver that, when used, leads to establishing a direct connection.
This seems to go against the idea of what PURPLE_PROXY_NONE setting
should mean and in its effect makes impossible for the user to fully
disable proxy in libpurple.

Instead of the default resolver, create a GSimpleProxyResolver with
an empty default_proxy, which will always return 'direct://' proxy.